8289231: Move java_lang_VirtualThread to continuationJavaClasses.hpp
