+++ src/hotspot/share/prims/jvmtiExtensions.cpp
@@ -27,6 +27,7 @@
 #include "prims/jvmtiExport.hpp"
 #include "prims/jvmtiExtensions.hpp"
 #include "prims/jvmtiThreadState.inline.hpp"
+#include "runtime/continuationJavaClasses.hpp"
 #include "runtime/handles.inline.hpp"
 #include "runtime/interfaceSupport.inline.hpp"
 #include "runtime/jniHandles.inline.hpp"
