+++ src/hotspot/share/runtime/continuationJavaClasses.cpp
@@ -25,9 +25,100 @@
 #include "precompiled.hpp"
 #include "classfile/javaClasses.hpp"
 #include "classfile/javaClassesImpl.hpp"
+#include "classfile/javaThreadStatus.hpp"
 #include "classfile/vmClasses.hpp"
 #include "classfile/vmSymbols.hpp"
-#include "runtime/continuationJavaClasses.hpp"
+#include "runtime/continuationJavaClasses.inline.hpp"
+
+// java_lang_VirtualThread
+
+int java_lang_VirtualThread::static_notify_jvmti_events_offset;
+int java_lang_VirtualThread::static_vthread_scope_offset;
+int java_lang_VirtualThread::_carrierThread_offset;
+int java_lang_VirtualThread::_continuation_offset;
+int java_lang_VirtualThread::_state_offset;
+
+#define VTHREAD_FIELDS_DO(macro) \
+  macro(static_notify_jvmti_events_offset, k, "notifyJvmtiEvents",  bool_signature,              true);  \
+  macro(static_vthread_scope_offset,       k, "VTHREAD_SCOPE",      continuationscope_signature, true);  \
+  macro(_carrierThread_offset,             k, "carrierThread",      thread_signature,            false); \
+  macro(_continuation_offset,              k, "cont",               continuation_signature,      false); \
+  macro(_state_offset,                     k, "state",              int_signature,               false)
+
+static bool vthread_notify_jvmti_events = JNI_FALSE;
+
+void java_lang_VirtualThread::compute_offsets() {
+  InstanceKlass* k = vmClasses::VirtualThread_klass();
+  VTHREAD_FIELDS_DO(FIELD_COMPUTE_OFFSET);
+}
+
+void java_lang_VirtualThread::init_static_notify_jvmti_events() {
+  if (vthread_notify_jvmti_events) {
+    InstanceKlass* ik = vmClasses::VirtualThread_klass();
+    oop base = ik->static_field_base_raw();
+    base->release_bool_field_put(static_notify_jvmti_events_offset, vthread_notify_jvmti_events);
+  }
+}
+
+bool java_lang_VirtualThread::is_instance(oop obj) {
+  return obj != NULL && is_subclass(obj->klass());
+}
+
+oop java_lang_VirtualThread::carrier_thread(oop vthread) {
+  oop thread = vthread->obj_field(_carrierThread_offset);
+  return thread;
+}
+
+oop java_lang_VirtualThread::continuation(oop vthread) {
+  oop cont = vthread->obj_field(_continuation_offset);
+  return cont;
+}
+
+int java_lang_VirtualThread::state(oop vthread) {
+  return vthread->int_field_acquire(_state_offset);
+}
+
+JavaThreadStatus java_lang_VirtualThread::map_state_to_thread_status(int state) {
+  JavaThreadStatus status = JavaThreadStatus::NEW;
+  switch (state) {
+    case NEW :
+      status = JavaThreadStatus::NEW;
+      break;
+    case STARTED :
+    case RUNNABLE :
+    case RUNNABLE_SUSPENDED :
+    case RUNNING :
+    case PARKING :
+    case YIELDING :
+      status = JavaThreadStatus::RUNNABLE;
+      break;
+    case PARKED :
+    case PARKED_SUSPENDED :
+    case PINNED :
+      status = JavaThreadStatus::PARKED;
+      break;
+    case TERMINATED :
+      status = JavaThreadStatus::TERMINATED;
+      break;
+    default:
+      ShouldNotReachHere();
+  }
+  return status;
+}
+
+#if INCLUDE_CDS
+void java_lang_VirtualThread::serialize_offsets(SerializeClosure* f) {
+   VTHREAD_FIELDS_DO(FIELD_SERIALIZE_OFFSET);
+}
+#endif
+
+bool java_lang_VirtualThread::notify_jvmti_events() {
+  return vthread_notify_jvmti_events == JNI_TRUE;
+}
+
+void java_lang_VirtualThread::set_notify_jvmti_events(bool enable) {
+  vthread_notify_jvmti_events = enable;
+}
 
 // Support for jdk.internal.vm.ContinuationScope
